refactor: privatization 2
