+++ cognite/neat/graph/stores/__init__.py
@@ -1,5 +1,5 @@
 from ._base import NeatGraphStoreBase, drop_graph_store_storage
-from .configuration import RdfStoreConfig, RdfStoreType
-from .rdf_to_graph import rdf_file_to_graph
+from ._configuration import RdfStoreConfig, RdfStoreType
+from ._rdf_to_graph import rdf_file_to_graph
 
 __all__ = ["NeatGraphStoreBase", "drop_graph_store_storage", "RdfStoreType", "RdfStoreConfig", "rdf_file_to_graph"]
