+++ cognite/neat/graph/stores/_configuration.py
